Fix GetUplinkRepresentor - Detecting incorrect Uplink Representor in new kernel
Supprt new kernel change [1] which links the PF and VF representors
to their parent PCI device similar to existing uplink representor
netdevice.
This leads that the PF and VF representors are exiting in
the VF_PCI_ADDR_Path/physfn/net which can cause selecting the wrong
uplink represenotr for VF.
The issue is solved by checking the phys_port_name
for which is in format 'p<port_num>' format for the PF and
'pf<pf_index>vf<vf_index>' for the VF representor.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net.git/commit/?id=123f0f53dd64b67e34142485fe866a8a581f12f1