Refactoring - Change Smart-NIC to DPU
Data Processing Units are a new breed of Smart-NICs
A key difference is having a mulitcore processor able
to run commodity software as part of the SOC.

Some operations performed on DPUs are not supported
by standard Smart-NICs, properly rename relevant
methods to reflect that.

Note: As there has not been a release with the old
names, it is OK at this point to change naming of
user facing methods.

Signed-off-by: Adrian Chiris <adrianc@nvidia.com>